src/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "arena_models"
version = "0.1.0"
description = "TODO"
authors = [
    { name = "kuro", email = "giang.nht108201@gmail.com" },
    { name = "Vova Shcherbyna", email = "dev@voshch.dev" },
]
dependencies = [
    "gdown",
    "bpy~=4.5.3",
    "chromadb~=1.0.21",
    "numpy~=1.26.4",
    "scipy~=1.15.3",
    "spacy~=3.8.7",
    "en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl",
    "pyyaml~=6.0.2",
    "mathutils~=3.3.0",
    "attrs~=25.3.0",
    "cattrs~=25.2.0",
    "typer~=0.19.1",
]
requires-python = "==3.11.*"

[project.urls]
Homepage = "https://github.com/arena-models/arena-models"
Repository = "https://github.com/arena-models/arena-models"


[tool.setuptools.packages.find]
where = ["."]
include = ["arena_models*"]